Basic Animation over Network
Added two basic animations, Idle and Walking. These are just for testing purposes. I also created an animation controller for testing. Added to animation scripts to changed between the animations, as well as in the Movement Controller to align animation with the input. Also flipped the animation when going different directions, actually done through setting the localScale to -1. Also got rid of an issue with the Network Manager Script, under AddGamePlayer. When players are added, this is where it determines their role, but only the server should be able to do this. The issue was that everyone on every instance was setting this themselves, so I added some logic to check if it is the server instance that is requesting.