Added Item Pickup
I created equipment to pickup: an axe, a pickaxe, and a fishing pole.

I created a gameobject for the axe. When you collide with it you can press P to pick it up. Once equipped, you can click or press space to use.

I tried to setup a simple animation using rotations. It does not currently work well.
It also does not track with the player well either alone or across multiple clients.