Added a joint storage system for materials.
Created an object where items can be stored to be added to the town inventory count. When you walk on top of one of these Storage areas, a simple UI pops up that shows the total amount of inventory that is available. I have not tested Creating more storage areas, I only tested one.