Updated Mirror Networking Asset. Restructured Items.
I updated Mirror to the latest version.

I added a Constants script to hold all the constants in. Right now it only has a list of all the items types that exist.

I created scriptable objects to hold the data for the Items.
I restructured the Equip able Items, Materials, and Storage classes to match the new Constant Item Types.